added functionality to complete RESTWS and fix package structure so @componentScan can find controller
