+++ src/main/java/com/collabera/InGiven/InGivenWebApplication.java
@@ -1,5 +1,6 @@
 package com.collabera.InGiven;
 
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
