+++ src/main/java/com/collabera/InGiven/controller/OtherITemsController.java
@@ -1,4 +1,4 @@
-package com.collabera.restapi.restcontroller;
+package com.collabera.InGiven.controller;
 
 
 
@@ -18,8 +18,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.collabera.restapi.model.ItemDTO;
-import com.collabera.restapi.service.OtherItemsService;
+import com.collabera.InGiven.model.ItemDTO;
+import com.collabera.InGiven.service.OtherItemsService;
 
 @RestController
 @RequestMapping("/api")
