+++ src/main/java/com/collabera/InGiven/model/ItemMapper.java
@@ -1,4 +1,4 @@
-package com.collabera.restapi.model;
+package com.collabera.InGiven.model;
 
 import org.springframework.stereotype.Component;
