+++ src/main/java/com/collabera/InGiven/model/ToyDTO.java
@@ -1,4 +1,4 @@
-package com.collabera.restapi.model;
+package com.collabera.InGiven.model;
 
 public class ToyDTO {
 	private long id;
