+++ src/main/java/com/collabera/InGiven/model/ToyMapper.java
@@ -1,5 +1,8 @@
-package com.collabera.restapi.model;
+package com.collabera.InGiven.model;
 
+import org.springframework.stereotype.Component;
+
+@Component
 public class ToyMapper { // should be static methods in their respective classes
 	public Toy toModel(ToyDTO in) {
 		Toy out = new Toy();
