+++ src/main/java/com/collabera/InGiven/repository/BookRepository.java
@@ -1,8 +1,9 @@
-package com.collabera.restapi.Repository;
+package com.collabera.InGiven.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.collabera.restapi.model.Book;
+
+import com.collabera.InGiven.model.Book;
 
 @Repository
 public interface BookRepository extends JpaRepository<Book, Long> {
