+++ src/main/java/com/collabera/InGiven/repository/ClothesRepository.java
@@ -1,9 +1,9 @@
-package com.collabera.restapi.Repository;
+package com.collabera.InGiven.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.collabera.restapi.model.Clothes;
+import com.collabera.InGiven.model.Clothes;
 
 @Repository
 public interface ClothesRepository extends JpaRepository<Clothes,Long> {
