+++ src/main/java/com/collabera/InGiven/repository/ToyRepository.java
@@ -1,9 +1,9 @@
-package com.collabera.restapi.Repository;
+package com.collabera.InGiven.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.collabera.restapi.model.Toy;
+import com.collabera.InGiven.model.Toy;
 
 @Repository
 public interface ToyRepository extends JpaRepository<Toy, Long>{}
