+++ src/main/java/com/collabera/InGiven/service/ToyService.java
@@ -1,4 +1,4 @@
-package com.collabera.restapi.service;
+package com.collabera.InGiven.service;
 
 import java.util.Optional;
 
@@ -10,10 +10,10 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.collabera.restapi.Repository.ToyRepository;
-import com.collabera.restapi.model.Toy;
-import com.collabera.restapi.model.ToyDTO;
-import com.collabera.restapi.model.ToyMapper;
+import com.collabera.InGiven.model.Toy;
+import com.collabera.InGiven.model.ToyDTO;
+import com.collabera.InGiven.model.ToyMapper;
+import com.collabera.InGiven.repository.ToyRepository;
 
 @Service
 public class ToyService {
