+++ src/main/resources/application.yml
@@ -15,5 +15,3 @@ spring:
 
 server:
   port: 8080
-
-
